OMS OrderController Code Review Fix
